Split Memo and PersonDatabase into their own files
